Fix data type

+++ database/migrations/2025_08_11_085645_add_master_collection_stats_to_masters.php
@@ -17,12 +17,12 @@ return new class extends Migration
     {
         Schema::table('master_shop_stats', function (Blueprint $table) {
              foreach (CollectionProductsStatusEnum::cases() as $state) {
-                $table->unsignedSmallInteger('number_collections_products_status_'.$state->snake())->default(0);
+                $table->unsignedInteger('number_collections_products_status_'.$state->snake())->default(0);
             }
         });
         Schema::table('master_product_category_stats', function (Blueprint $table) {
             foreach (CollectionStateEnum::cases() as $state) {
-                $table->unsignedSmallInteger('number_collections_state_'.$state->snake())->default(0);
+                $table->unsignedInteger('number_collections_state_'.$state->snake())->default(0);
             }
         });
     }
